Remove switch statement and rename routes to views
