+++ src/App.js
@@ -44,7 +44,7 @@ import initIntercomSettings from "./utils/initIntercomSettings";
 import { initI18n } from "./utils/i18nSettings";
 import { LANTERN_TENANTS } from "./RootStore/TenantStore/lanternTenants";
 import { CORE_TENANTS } from "./RootStore/TenantStore/coreTenants";
-import { CORE_PATHS } from "./core/routes";
+import { CORE_PATHS } from "./core/views";
 import AuthWall from "./AuthWall";
 import ErrorBoundary from "./components/ErrorBoundary";
